track 'snap' branch instead of 'hpsc'
It won't be practically feasible to maintain the invariant that if a
commit is in 'hpsc' branch of the child repo then it is in a snapshot in
this repo; because others may choose to commit into hpsc without making
a snapshot. So, maintain this same invariant but w.r.t. 'snap' branch
instead.